pyproject.toml
[project]
name = "liquidationheatmap"
version = "0.1.0"
description = "Claude Code project: LiquidationHeatmap"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.9.0",
    "duckdb>=0.9.0",
    "fastapi>=0.104.0",
    "redis>=5.0.0",
    "pydantic>=2.5.0",
    "plotly>=5.17.0",
    "prometheus-client>=0.20.0",
    "uvicorn>=0.24.0",
    "websockets>=12.0",
    "pandas>=2.1.0",
    "rich>=14.2.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "pyyaml>=6.0.3",
    "apscheduler>=3.10.0",
    "httpx>=0.25.0",
    "cachetools>=5.3.0",
    "scikit-learn>=1.3.0",
    "pytesseract>=0.3.13",
    "pillow>=12.0.0",
    "opencv-python>=4.11.0.86",
    "tqdm>=4.67.1",
    "pyotp>=2.9.0",
    "pycryptodome>=3.23.0",
    "zstandard>=0.22.0",
    "msgpack>=1.1.2",
    "msgspec>=0.20.0",
    "questdb>=4.1.0",
    "psycopg2-binary>=2.9.11",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-benchmark>=4.0.0",
    "hypothesis>=6.90.0",
    "mypy>=1.7.0",
    "ruff>=0.1.0",
    "pyarrow>=14.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
markers = [
    "integration: tests requiring real external data or services",
]
# addopts = "--cov=src --cov-report=term-missing --cov-report=html"

[tool.ruff]
line-length = 100
target-version = "py311"
extend-exclude = [
    ".claude",
    "archive",
    "docs",
    "examples",
    "frontend",
    "scripts",
    "tests",
    "ingest_full_history_n8n.py",
    "ingest_full_history_safe.py",
    "src/api",
    "src/clustering",
    "src/config",
    "src/db",
    "src/exchanges",
    "src/models",
    "src/services",
    "src/validation",
]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"src/liquidationheatmap/api/main.py" = ["E402"]

[tool.mutmut]
paths_to_mutate = [
    "src/liquidationheatmap/ingestion/gap_fill.py",
    "src/liquidationheatmap/ingestion/db_service.py",
    "src/liquidationheatmap/api/main.py"
]
tests_dir = "tests/"
backup = false
runner = "python3 -m pytest -x"
dict_synonyms = ["Struct", "dict"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "hypothesis>=6.148.2",
    "playwright>=1.58.0",
]
